Refactor account class to account actions
The account class in the `banking` package has been renamed to `AccountActions`. This change reflects a better and more accurate representation of the functionalities provided by the class. The `AccountActions` class contains a `balance` property and methods for depositing and withdrawing amounts.

The usage of the `Account` class has been replaced with `AccountActions` in the `Main.kt` file. The instance of `AccountActions` is created using the `accActions` variable. The interactions with the account, such as deposit, withdraw, and display balance, have been updated to use `accActions` instead of `Account`.

Additionally, the `User` class in the `banking` package has been removed. This class was not referenced or utilized in the codebase.

A new `todo.txt` file has been added. This file contains a todo list related to fixing some issues or implementing certain features. The items in the todo list include clarifying the available operations and handling invalid inputs for deposit amounts.

Please review and merge these changes accordingly.